pyproject.toml
[project]
name = "dots-ocr-client"
version = "0.1.0"
description = "A Python client library for dots.ocr VLLM service"
readme = "README.md"
authors = [
    { name = "sljeff", email = "kindjeff.com@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "openai>=1.99.1",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pypdfium2>=4.30.0",
    "replicate>=1.0.7",
    "requests>=2.32.4",
]

[project.scripts]
dots-ocr-client = "dots_ocr_client.parser:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dots_ocr_client"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
